# Make jsonObfuscator concurrency-safe by internal state isolation