--- a/pkg/obfuscate/json_test.go
+++ b/pkg/obfuscate/json_test.go
@@ -152,40 +152,41 @@ func FuzzObfuscateJSON(f *testing.F) {
 	})
 }
 
-func TestObfuscateJSONConcurrently(t *testing.T) {
+func BenchmarkObfuscateJSONConcurrently(b *testing.B) {
 	var (
 		cfg = &JSONConfig{
 			KeepValues: []string{"company_wallet_configuration_id"},
 		}
 		o        = newJSONObfuscator(cfg, &Obfuscator{})
-		rs       = make(chan string, 1)
 		expected = "{\"email\":\"?\",\"company_wallet_configuration_id\":1}"
 	)
 
-	// Concurrency issues were reproduced at 100 times, but we leave this at 100K
-	// to stress test the obfuscator.
-	const times = 100000
-	var wg sync.WaitGroup
-	for i := 0; i < times; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			in := "{\"email\":\"dev@datadoghq.com\",\"company_wallet_configuration_id\":1}"
-			rs <- obfuscateJSONString(in, o)
-		}()
-	}
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		rs := make(chan string, 1)
 
-	go func() {
-		wg.Wait()
-		close(rs)
-	}()
-
-	for i := 0; i < times; i++ {
-		v := <-rs
-		if v == expected {
-			continue
+		// Concurrency issues were reproduced at 100 times, but we leave this at 100K
+		// to stress test the obfuscator.
+		const times = 100000
+		var wg sync.WaitGroup
+		for i := 0; i < times; i++ {
+			wg.Add(1)
+			go func() {
+				defer wg.Done()
+				in := "{\"email\":\"dev@datadoghq.com\",\"company_wallet_configuration_id\":1}"
+				rs <- obfuscateJSONString(in, o)
+			}()
 		}
 
-		t.Fatalf("mangled payload after obfuscation: %q", v)
+		for i := 0; i < times; i++ {
+			v := <-rs
+			if v == expected {
+				continue
+			}
+			b.Fatalf("mangled payload after obfuscation: %q", v)
+		}
+
+		wg.Wait()
+		close(rs)
 	}
 }