--- a/pkg/trace/agent/agent.go
+++ b/pkg/trace/agent/agent.go
@@ -201,7 +201,6 @@ func (a *Agent) loop() {
 		a.EventProcessor,
 		a.OTLPReceiver,
 		a.obfuscator,
-		a.obfuscator,
 		a.cardObfuscator,
 		a.DebugServer,
 	} {