… pkg/obfuscate/testdata/json_tests.xml