@darccio 2023-09-22
I measured the impact of allocating the new internal state struct and I found that it was significant (in bytes and nanoseconds scale):

main:

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1381185               855.3 ns/op           384 B/op          7 allocs/op
```

This PR:

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1000000              1053 ns/op             576 B/op         13 allocs/op
BenchmarkObfuscateJSON/obfuscate.mongo.json.keep_values-10                               1979707               606.8 ns/op           400 B/op         11 allocs/op
```

I applied two optimizations ([`out.Grow()`](https://github.com/DataDog/datadog-agent/pull/19643/files#diff-7ce08f4ef3a4c33487f631696e3fe94835170c85781693113325c558be655abbR109) and [`bytes.Trim()` instead of `strings.Trim()`](https://github.com/DataDog/datadog-agent/pull/19643/files#diff-7ce08f4ef3a4c33487f631696e3fe94835170c85781693113325c558be655abbR172)) that improved performance, setting it closer to the original implementation:

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1268534               943.5 ns/op           440 B/op          9 allocs/op
BenchmarkObfuscateJSON/obfuscate.mongo.json.keep_values-10                               2123896               549.2 ns/op           312 B/op          7 allocs/op
```

PS: if we apply the optimizations to main, we see that the performance impact of the fix would be proportionally to the baseline, doubling allocation. At least it has been mitigated and set almost to the original implementation's performance.

```
BenchmarkObfuscateJSON/elasticsearch.body.1-10                                           1585716               754.8 ns/op           248 B/op          3 allocs/op
```