@darccio 2023-09-22
Optimizations and release notes splitted in separate PR #19682